+++ test/test_tmc_paper_dataset.cc
@@ -128,7 +128,7 @@ BOOST_AUTO_TEST_CASE( tmc_paper_dataset_evaluation ) {
     auto pEvaluator = std::make_shared<Evaluator>();
     std::pair<std::pair<double, double>, std::pair<double, double>> confusion_matrix_train, confusion_matrix_eval;
     double roc_err_train, roc_err_eval, err_rate_train, err_rate_eval;
-    unsigned int counter = 0;
+    size_t counter = 0;
     do {
         const auto pLP = pLayerPartitioningCreator->get_layer_partitioning();
         const auto pBSC = std::make_shared<BorderSystemCreator>();
@@ -195,7 +195,7 @@ BOOST_AUTO_TEST_CASE( tmc_paper_dataset_evaluation_sample_split ) {
     auto pEvaluator = std::make_shared<Evaluator>();
     std::pair<std::pair<double, double>, std::pair<double, double>> confusion_matrix_train, confusion_matrix_eval;
     double roc_err_train, roc_err_eval, err_rate_train, err_rate_eval;
-    unsigned int counter = 0;
+    size_t counter = 0;
     do {
         const auto pLP = pLayerPartitioningCreator->get_layer_partitioning();
         const auto pBSC = std::make_shared<BorderSystemCreator>();
